Update to teleport order
This is a proposed fix for some contraptions that seem to slow down after teleporting.

With this change, the angle and velocity of each prop is logged before teleporting, rather than logging/teleporting/setting each prop individually. This is a more expensive method than the previous one, but is designed to mitigate any angle/velocity mismatches that happen when teleporting to try and preserve the original  speed as much as possible.

Tested with prop vehicles/wac/simphys extensively.